Refactor, rename, move YouTubeVideoEmbed
